First set of working tests
